% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repsim_hsic.R
\name{repsim_hsic}
\alias{repsim_hsic}
\title{List of HSIC estimators}
\usage{
repsim_hsic()
}
\value{
A character vector of estimator names.
}
\description{
Returns a character vector of available HSIC estimators implemented in 
the package.
}
\references{
\insertRef{gretton_2005_MeasuringStatisticalDependence}{repsim}

\insertRef{song_2007_SupervisedFeatureSelection}{repsim}

\insertRef{lange_2023_DeepNetworksPaths}{repsim}
}
