% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repsim_kernels.R
\name{repsim_kernels}
\alias{repsim_kernels}
\title{List of kernel functions}
\usage{
repsim_kernels()
}
\value{
A character vector of kernel names.
}
\description{
Returns a character vector of available kernel methods implemented in the package.
}
